ex00: double -> float

+++ ex00/includes/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: nseon <nseon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/26 16:52:44 by nseon             #+#    #+#             */
-/*   Updated: 2026/01/30 14:31:50 by nseon            ###   ########.fr       */
+/*   Updated: 2026/02/20 11:14:48 by nseon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 
 class BitcoinExchange {
 	private:
-		std::map<std::string, double> _map;
+		std::map<std::string, float> _map;
 	public:
 		BitcoinExchange();
 		BitcoinExchange(char *input_file);
